==> 02_covariate_analysis/01_mediator_trait_correlation.R <==
########################################################
# Libraries
########################################################
library(dplyr)
library(data.table)



########################################################
# STEP 1: Load data
########################################################

# Load continuous phenotype data (non-transformed; identifier + 1 column/trait); 
# Sex-specific traits originate from sex-specific data files and are appended to the main dataframe
pheno_All <- as.data.frame(fread("pheno_continuous_WB_raw_All.txt.gz"))
pheno_M <- as.data.frame(fread("pheno_continuous_WB_raw_M.txt.gz"))
pheno_F <- as.data.frame(fread("pheno_continuous_WB_raw_F.txt.gz"))
pheno <- left_join(pheno_All, pheno_M[, names(pheno_M) %in% c("IID", "balding", "facial_hair")], by = "IID"); rm(pheno_All, pheno_M)
pheno <- left_join(pheno, pheno_F[, names(pheno_F) %in% c("IID", "menarche", "menopause", "birth_weight_first_child")], by = "IID"); rm(pheno_F)

# Load disease status data (identifier + 1 column/trait); 
# Sex-specific diseases originate from sex-specific data files and are appended to the main dataframe; 
# Data file is encoded so that cases are 1, controls are 0, and missing data NA
disease_All <- as.data.frame(fread("pheno_ICD10_All.txt.gz"))
disease_M <- as.data.frame(fread("pheno_ICD10_M.txt.gz"))
disease_F <- as.data.frame(fread("pheno_ICD10_F.txt.gz"))
disease <- left_join(disease_All, disease_M[, names(disease_M) %in% c("IID", "balding", "facial_hair")], by = "IID"); rm(disease_All, disease_M)
disease <- left_join(disease, disease_F[, names(disease_F) %in% c("IID", "BC", "OC", "endometriosis", "menstruation")], by = "IID"); rm(disease_F)

# Load potential mediators on a raw scale (identifier + BMI/EA/height)
bmi <- as.data.frame(fread("pheno_continuous_WB_raw_All.txt", select = c(1,6))) # Load single column from the above file
ea <- as.data.frame(fread("EA_WB_raw_All.txt")) # Generated by script/01_PheWAS/01_EA_phenotype_definition.R
height <- as.data.frame(fread("pheno_continuous_WB_raw_All.txt", select = c(1,2))) # Load single column from the above file



########################################################
# STEP 2: Correlation with continuous traits
########################################################

### BMI ################################################

# Create an empty dataframe to store results
bmi_mediated_pheno <- data.frame(PHENO = names(pheno)[names(pheno) != "IID"], BETA = NA, P = NA, COR = NA)

# Loop over phenotypes
for (p in bmi_mediated_pheno$PHENO) {
  
  # Merge data & fit model
  df <- left_join(na.omit(pheno[, c("IID", p)]), bmi, by = "IID")[2:3]
  colnames(df) <- c("PHENO", "BMI") 
  fit <- lm(PHENO ~  BMI, data = df)
  
  # Fill in table
  bmi_mediated_pheno[which(bmi_mediated_pheno$PHENO == p), "BETA"] <- summary(fit)$coef[2,1]
  bmi_mediated_pheno[which(bmi_mediated_pheno$PHENO == p), "P"] <- summary(fit)$coef[2,4]
  bmi_mediated_pheno[which(bmi_mediated_pheno$PHENO == p), "COR"] <- cor(df$PHENO, df$BMI, use = "pairwise.complete.obs")

}; rm(p, df, fit)

# Save results
fwrite(bmi_mediated_pheno, "/data/correction_BMI/16p11.2_BP4-5_continuous_BMIassociation.txt", col.name = T, row.name = F, sep = "\t", quote = F)



### EA ################################################

# Create an empty dataframe to store results
ea_mediated_pheno <- data.frame(PHENO = names(pheno)[names(pheno) != "IID"], BETA = NA, P = NA, COR = NA)

# Loop over phenotypes
for (p in ea_mediated_pheno$PHENO) {
  
  # Merge data & fit model
  df <- left_join(na.omit(pheno[, c("IID", p)]), ea, by = "IID")[2:3]
  colnames(df) <- c("PHENO", "EA") 
  fit <- lm(PHENO ~  EA, data = df)
  
  # Fill in table
  ea_mediated_pheno[which(ea_mediated_pheno$PHENO == p), "BETA"] <- summary(fit)$coef[2,1]
  ea_mediated_pheno[which(ea_mediated_pheno$PHENO == p), "P"] <- summary(fit)$coef[2,4]
  ea_mediated_pheno[which(ea_mediated_pheno$PHENO == p), "COR"] <- cor(df$PHENO, df$EA, use = "pairwise.complete.obs")

  }; rm(p, df, fit)

# Save results
fwrite(ea_mediated_pheno, "/data/correction_EA/16p11.2_BP4-5_continuous_EAassociation.txt", col.name = T, row.name = F, sep = "\t", quote = F)



### HEIGHT ################################################

# Create an empty dataframe to store results
height_mediated_pheno <- data.frame(PHENO = names(pheno)[names(pheno) != "IID"], BETA = NA, P = NA, COR = NA)

# Loop over phenotypes
for (p in height_mediated_pheno$PHENO) {
  
  # Merge data & fit model
  df <- left_join(na.omit(pheno[, c("IID", p)]), height, by = "IID")[2:3]
  colnames(df) <- c("PHENO", "HEIGHT") 
  fit <- lm(PHENO ~  HEIGHT, data = df)
  
  # Fill in table
  height_mediated_pheno[which(height_mediated_pheno$PHENO == p), "BETA"] <- summary(fit)$coef[2,1]
  height_mediated_pheno[which(height_mediated_pheno$PHENO == p), "P"] <- summary(fit)$coef[2,4]
  height_mediated_pheno[which(height_mediated_pheno$PHENO == p), "COR"] <- cor(df$PHENO, df$HEIGHT, use = "pairwise.complete.obs")

}; rm(p, df, fit)

# Save results
fwrite(height_mediated_pheno, "/data/correction_height/16p11.2_BP4-5_continuous_HEIGHTassociation.txt", col.name = T, row.name = F, sep = "\t", quote = F)



########################################################
# STEP 3: Correlation with binary traits
########################################################

### BMI ################################################

# Create an empty dataframe to store results
bmi_mediated_disease <- data.frame(PHENO = names(disease)[names(disease) != "IID"], BETA = NA, P = NA, COR = NA)

# Loop over diseases
for (p in bmi_mediated_disease$PHENO) {
  
  # Merge data & fit model
  df <- left_join(na.omit(disease[, c("IID", p)]), bmi, by = "IID")[2:3]
  colnames(df) <- c("PHENO", "BMI") 
  fit <- logistf(PHENO ~ BMI, data = df, plconf = 2, maxit = 100, maxstep = 10)
  
  # Fill in table
  bmi_mediated_disease[which(bmi_mediated_disease$PHENO == p), "BETA"] <- fit$coefficients[2]
  bmi_mediated_disease[which(bmi_mediated_disease$PHENO == p), "P"] <- fit$prob[2]
  bmi_mediated_disease[which(bmi_mediated_disease$PHENO == p), "COR"] <- cor(df$PHENO, df$BMI, use = "pairwise.complete.obs")

}; rm(p, df, fit)

# Save results
fwrite(bmi_mediated_disease, "/data/correction_BMI/16p11.2_BP4-5_binary_BMIassociation.txt", col.name = T, row.name = F, sep = "\t", quote = F)



### EA ################################################

# Create an empty dataframe to store results
ea_mediated_disease <- data.frame(PHENO = names(disease)[names(disease) != "IID"], BETA = NA, P = NA, COR = NA)

# Loop over diseases
for (p in ea_mediated_disease$PHENO) {
  
  # Merge data & fit model
  df <- left_join(na.omit(disease[, c("IID", p)]), ea, by = "IID")[2:3]
  colnames(df) <- c("PHENO", "EA") 
  fit <- logistf(PHENO ~ EA, data = df, plconf = 2, maxit = 100, maxstep = 10)
  
  # Fill in table
  ea_mediated_disease[which(ea_mediated_disease$PHENO == p), "BETA"] <- fit$coefficients[2]
  ea_mediated_disease[which(ea_mediated_disease$PHENO == p), "P"] <- fit$prob[2]
  ea_mediated_disease[which(ea_mediated_disease$PHENO == p), "COR"] <- cor(df$PHENO, df$EA, use = "pairwise.complete.obs")
  
}; rm(p, df, fit)

# Save results
fwrite(ea_mediated_disease, "/data/correction_EA/16p11.2_BP4-5_binary_EAassociation.txt", col.name = T, row.name = F, sep = "\t", quote = F)



### HEIGHT ################################################

# Create an empty dataframe to store results
height_mediated_disease <- data.frame(PHENO = names(disease)[names(disease) != "IID"], BETA = NA, P = NA, COR = NA)

# Loop over diseases
for (p in height_mediated_disease$PHENO) {
  
  # Merge data & fit model
  df <- left_join(na.omit(disease[, c("IID", p)]), height, by = "IID")[2:3]
  colnames(df) <- c("PHENO", "HEIGHT") 
  fit <- logistf(PHENO ~ HEIGHT, data = df, plconf = 2, maxit = 100, maxstep = 10)
  
  # Fill in table
  height_mediated_disease[which(height_mediated_disease$PHENO == p), "BETA"] <- fit$coefficients[2]
  height_mediated_disease[which(height_mediated_disease$PHENO == p), "P"] <- fit$prob[2]
  height_mediated_disease[which(height_mediated_disease$PHENO == p), "COR"] <- cor(df$PHENO, df$HEIGHT, use = "pairwise.complete.obs")
  
}; rm(p, df, fit)

# Save results
fwrite(height_mediated_disease, "/data/correction_height/16p11.2_BP4-5_binary_HEIGHTassociation.txt", col.name = T, row.name = F, sep = "\t", quote = F)



########################################################
# STEP 4: Merge continuous/binary results & plot
########################################################

# Merge continuous and binary dataframes
bmi_mediated <- rbind(bmi_mediated_pheno, bmi_mediated_disease)
bmi_mediated$MEDIATOR <- "BMI"
ea_mediated <- rbind(ea_mediated_pheno, ea_mediated_disease)
ea_mediated$MEDIATOR <- "EA"
height_mediated <- rbind(height_mediated_pheno, height_mediated_disease)
height_mediated$MEDIATOR <- "HEIGHT"
df_mediated <- rbind(bmi_mediated, ea_mediated, height_mediated)
df_mediated$MEDIATOR <- factor(df_mediated$MEDIATOR, levels = c("BMI", "EA", "HEIGHT"), labels = c("BMI", "EA", "height")) # 3*117 = 351 pairs

# Save results 
fwrite(df_mediated, "/data/16p11.2_BP4-5_mediator_correlation.txt", col.name = T, row.name = F, sep = "\t", quote = F)

# Plot correlation
p_correlation <- ggplot(df_mediated, aes(x = COR)) +
                  # Points
                  geom_histogram(bins = 20, fill = "gray50", color = "gray20") +
                  # Facet
                  facet_grid(MEDIATOR ~ .) +
                  # Axis
                  coord_cartesian(xlim = c(-1,1)) +
                  # Layout
                  xlab("Pearson correlation") + ylab("Counts") +
                  theme_bw() +
                  theme(text = element_text(family = "Arial"),
                        axis.text.x = element_text(size = 12), 
                        axis.text.y = element_text(size = 10), 
                        axis.title.x = element_text(size = 15, margin = margin(t = 7, r = 0, b = 0, l = 0)),
                        axis.title.y = element_text(size = 15, margin = margin(t = 0, r = 7, b = 0, l = 0)),
                        legend.title = element_text(size = 11, face = "bold"), legend.text = element_text(size = 10),
                        legend.key.size = unit(1, 'lines'),
                        strip.text = element_text(size = 10, color = "#2B4050", face = "bold"),
                        strip.background = element_rect(fill = "#F4F4F4", color = "#2B4050"))
            

# Save plot        
showtext_auto()   
p_correlation
ggsave("/plot/mediator_trait_corelation.pdf", width = 5, height = 4.5)
showtext_auto(FALSE)
