==> 02_covariate_analysis/02_PheWAS_covariate_corrected.R <==
########################################################
# Libraries
########################################################
library(dplyr)
library(data.table)
library(logistf)



########################################################
# STEP 1: Load data
########################################################

# Load sample file (vector with identifiers, with header "eid")
samples <- as.data.frame(fread("samples_white_british_All.txt"))

# Load CNV file and subset samples (output from CNV calling pipeline, as described in Auwerx et al., 2022 (AJHG))
cnvs <- as.data.frame(fread("ukb_cnv_global.gz"))
cnvs$Sample_Name <- sub("_.*", "", cnvs$Sample_Name)
cnvs <- cnvs[cnvs$Sample_Name %in% samples$eid, ]

# Load continuous phenotype data (inverse normal transformed + covariate-corrected; as described in Auwerx et al., 2022 (AJHG); identifier + 1 column/trait); 
# Sex-specific traits originate from sex-specific data files and are appended to the main dataframe
pheno_All <- as.data.frame(fread("pheno_continuous_WB_INT_age_age2_sex_batch_PCs_All.txt.gz"))
pheno_M <- as.data.frame(fread("pheno_continuous_WB_INT_age_age2_batch_PCs_M.txt.gz"))
pheno_F <- as.data.frame(fread("pheno_continuous_WB_INT_age_age2_batch_PCs_F.txt.gz"))
pheno <- left_join(pheno_All, pheno_M[, names(pheno_M) %in% c("IID", "balding", "facial_hair")], by = "IID"); rm(pheno_All, pheno_M)
pheno <- left_join(pheno, pheno_F[, names(pheno_F) %in% c("IID", "menarche", "menopause", "birth_weight_first_child")], by = "IID"); rm(pheno_F)

# Load disease status data (identifier + 1 column/trait); 
# Sex-specific diseases originate from sex-specific data files and are appended to the main dataframe; 
# Data file is encoded so that cases are 1, controls are 0, and missing data NA
disease_All <- as.data.frame(fread("pheno_ICD10_All.txt.gz"))
disease_M <- as.data.frame(fread("pheno_ICD10_M.txt.gz"))
disease_F <- as.data.frame(fread("pheno_ICD10_F.txt.gz"))
disease <- left_join(disease_All, disease_M[, names(disease_M) %in% c("IID", "balding", "facial_hair")], by = "IID"); rm(disease_All, disease_M)
disease <- left_join(disease, disease_F[, names(disease_F) %in% c("IID", "BC", "OC", "endometriosis", "menstruation")], by = "IID"); rm(disease_F)

# Load covariate data (identifier + 1 column/covariate, including age, sex, array, and PC1-40)
cov <- as.data.frame(fread("covariates.txt"))

# Load potential mediators on a raw scale (identifier + BMI/EA/height).
bmi <- as.data.frame(fread("pheno_continuous_WB_raw_All.txt", select = c(1,6))) # Same raw phenotype file as used in script/02_mediator_trait_correlation.R, but single column is loaded
ea <- as.data.frame(fread("EA_WB_raw_All.txt")) # Generated by script/01_PheWAS/01_EA_phenotype_definition.R
height <- as.data.frame(fread("pheno_continuous_WB_raw_All.txt", select = c(1,2))) # Same raw phenotype file as used in script/02_mediator_trait_correlation.R, but single column is loaded

# Load mediator-trait correlations generated in script/02_covariate_analysis/01_mediator_trait_correlation.R; 
# Filter for significant asociation (p < 0.05/351) & Pearson correlation < 0.7
df_cor <- as.data.frame(fread("/data/16p11.2_BP4-5_mediator_correlation.txt"))
df_cor <- df_cor[which(df_cor$P < 0.05/351), ]
df_cor <- df_cor[which(abs(df_cor$COR) < 0.7), ]



########################################################
# STEP 2: Generate genotype table
########################################################

# Identify CNV carriers according to the following criteria; Distinguish deletion and duplication carriers;
# Start: 29.4-29.8 Mb; End: 30.05-30.4 Mb; |QS| > 0.5 
cnvs <- cnvs[which(cnvs$Chromosome == 16 & cnvs$Start_Position_bp >= 29400000 & cnvs$Start_Position_bp <= 29800000 & cnvs$End_Position_bp >= 30050000 & cnvs$End_Position_bp <= 30400000 & abs(cnvs$Quality_Score) >= 0.5), ]
dup_carriers <- cnvs[which(cnvs$Copy_Number > 2), "Sample_Name"]
del_carriers <- cnvs[which(cnvs$Copy_Number < 2), "Sample_Name"]

# Create an empty data frame to store CNV genotype encoding to 4 models
cnv_geno <- data.frame(IID = samples$eid, DUP = 0, DEL = 0, M = 0, U = 0)

# Duplication-only model encoding
cnv_geno[cnv_geno$IID %in% del_carriers, "DUP"] <- NA
cnv_geno[cnv_geno$IID %in% dup_carriers, "DUP"] <- 1

# Deletion-only  model encoding
cnv_geno[cnv_geno$IID %in% dup_carriers, "DEL"] <- NA
cnv_geno[cnv_geno$IID %in% del_carriers, "DEL"] <- 1

# Mirror  model encoding
cnv_geno[cnv_geno$IID %in% del_carriers, "M"] <- -1
cnv_geno[cnv_geno$IID %in% dup_carriers, "M"] <- 1

# U-shape  model encoding
cnv_geno[cnv_geno$IID %in% c(dup_carriers, del_carriers), "U"] <- 1



########################################################
# STEP 3: Continuous traits
########################################################

### BMI ################################################

# Create a dataframe to store results
pheno_result_bmi <- df_cor[which(df_cor$MEDIATOR == "BMI"), "PHENO"]
pheno_result_bmi <- pheno_result_bmi[pheno_result_bmi %in% colnames(pheno)[-1]]
pheno_result_bmi <- data.frame(PHENO = rep(pheno_result_bmi, each = 4),
                               MODEL = rep(c("DUP", "DEL", "M", "U"), length(pheno_result_bmi)),
                               MEDIATOR = "BMI")

# Loop over phenotypes
for (p in unique(pheno_result_bmi$PHENO)) {
  
  print(paste0("Analyzing: ", p))
  
  # Loop over models
  for (m in c("DUP", "DEL", "M", "U")) {
    
    print(m)
    
    # Subset data (phenotype, model of interest, covariate)
    df <- left_join(na.omit(pheno[, c("IID", p)]), cnv_geno[, c("IID", m)], by = "IID")
    df <- left_join(df, bmi, by = "IID")[2:4]
    colnames(df) <- c("PHENO", "CNV", "BMI") 
    
    # Fit model
    fit <- lm(PHENO ~ CNV + BMI, data = df)
    
    # Fill in summary table
    pheno_result_bmi[which(pheno_result_bmi$PHENO == p & pheno_result_bmi$MODEL == m), "BETA"] <- summary(fit)$coeff[2,1]
    pheno_result_bmi[which(pheno_result_bmi$PHENO == p & pheno_result_bmi$MODEL == m), "SE"] <- summary(fit)$coeff[2,2]
    pheno_result_bmi[which(pheno_result_bmi$PHENO == p & pheno_result_bmi$MODEL == m), "P"] <- summary(fit)$coeff[2,4]
    pheno_result_bmi[which(pheno_result_bmi$PHENO == p & pheno_result_bmi$MODEL == m), "L95"] <- summary(fit)$coeff[2,1] - 1.96*summary(fit)$coeff[2,2]
    pheno_result_bmi[which(pheno_result_bmi$PHENO == p & pheno_result_bmi$MODEL == m), "U95"] <- summary(fit)$coeff[2,1] + 1.96*summary(fit)$coeff[2,2]
    
  }
  
}
rm(p, m, df, fit)

# Save data 
fwrite(pheno_result_bmi, "/data/correction_BMI/16p11.2_BP4-5_continuousPheWAS_BMIcorrected.txt", col.name = T, row.name = F, sep = "\t", quote = F)



### EA ################################################

# Create a dataframe to store results
pheno_result_ea <- df_cor[which(df_cor$MEDIATOR == "EA"), "PHENO"]
pheno_result_ea <- pheno_result_ea[pheno_result_ea %in% colnames(pheno)[-1]]
pheno_result_ea <- data.frame(PHENO = rep(pheno_result_ea, each = 4),
                               MODEL = rep(c("DUP", "DEL", "M", "U"), length(pheno_result_ea)),
                               MEDIATOR = "EA")

# Loop over phenotypes
for (p in unique(pheno_result_ea$PHENO)) {
  
  print(paste0("Analyzing: ", p))
  
  # Loop over models
  for (m in c("DUP", "DEL", "M", "U")) {
    
    print(m)
    
    # Subset data (phenotype, model of interest, covariate)
    df <- left_join(na.omit(pheno[, c("IID", p)]), cnv_geno[, c("IID", m)], by = "IID")
    df <- left_join(df, ea, by = "IID")[2:4]
    colnames(df) <- c("PHENO", "CNV", "EA") 
    
    # Fit model
    fit <- lm(PHENO ~ CNV + EA, data = df)
    
    # Fill in summary table
    pheno_result_ea[which(pheno_result_ea$PHENO == p & pheno_result_ea$MODEL == m), "BETA"] <- summary(fit)$coeff[2,1]
    pheno_result_ea[which(pheno_result_ea$PHENO == p & pheno_result_ea$MODEL == m), "SE"] <- summary(fit)$coeff[2,2]
    pheno_result_ea[which(pheno_result_ea$PHENO == p & pheno_result_ea$MODEL == m), "P"] <- summary(fit)$coeff[2,4]
    pheno_result_ea[which(pheno_result_ea$PHENO == p & pheno_result_ea$MODEL == m), "L95"] <- summary(fit)$coeff[2,1] - 1.96*summary(fit)$coeff[2,2]
    pheno_result_ea[which(pheno_result_ea$PHENO == p & pheno_result_ea$MODEL == m), "U95"] <- summary(fit)$coeff[2,1] + 1.96*summary(fit)$coeff[2,2]
    
  }
  
}
rm(p, m, df, fit)

# Save data
fwrite(pheno_result_ea, "/data/correction_EA/16p11.2_BP4-5_continuousPheWAS_EAcorrected.txt", col.name = T, row.name = F, sep = "\t", quote = F)



### HEIGHT ################################################

# Create a dataframe to store results
pheno_result_height <- df_cor[which(df_cor$MEDIATOR == "height"), "PHENO"]
pheno_result_height <- pheno_result_height[pheno_result_height %in% colnames(pheno)[-1]]
pheno_result_height <- data.frame(PHENO = rep(pheno_result_height, each = 4),
                               MODEL = rep(c("DUP", "DEL", "M", "U"), length(pheno_result_height)),
                               MEDIATOR = "height")

# Loop over phenotypes
for (p in unique(pheno_result_height$PHENO)) {
  
  print(paste0("Analyzing: ", p))
  
  # Loop over models
  for (m in c("DUP", "DEL", "M", "U")) {
    
    print(m)
    
    # Subset data (phenotype, model of interest, covariate)
    df <- left_join(na.omit(pheno[, c("IID", p)]), cnv_geno[, c("IID", m)], by = "IID")
    df <- left_join(df, height, by = "IID")[2:4]
    colnames(df) <- c("PHENO", "CNV", "height") 
    
    # Fit model
    fit <- lm(PHENO ~ CNV + height, data = df)
    
    # Fill in summary table
    pheno_result_height[which(pheno_result_height$PHENO == p & pheno_result_height$MODEL == m), "BETA"] <- summary(fit)$coeff[2,1]
    pheno_result_height[which(pheno_result_height$PHENO == p & pheno_result_height$MODEL == m), "SE"] <- summary(fit)$coeff[2,2]
    pheno_result_height[which(pheno_result_height$PHENO == p & pheno_result_height$MODEL == m), "P"] <- summary(fit)$coeff[2,4]
    pheno_result_height[which(pheno_result_height$PHENO == p & pheno_result_height$MODEL == m), "L95"] <- summary(fit)$coeff[2,1] - 1.96*summary(fit)$coeff[2,2]
    pheno_result_height[which(pheno_result_height$PHENO == p & pheno_result_height$MODEL == m), "U95"] <- summary(fit)$coeff[2,1] + 1.96*summary(fit)$coeff[2,2]
    
  }
  
}
rm(p, m, df, fit)

# Save data
fwrite(pheno_result_height, "/data/correction_height/16p11.2_BP4-5_continuousPheWAS_HEIGHTcorrected.txt", col.name = T, row.name = F, sep = "\t", quote = F)



########################################################
# STEP 4: Disease
########################################################

### BMI ################################################

# Create a dataframe to store results
diseases_result_bmi <- df_cor[which(df_cor$MEDIATOR == "BMI"), "PHENO"]
diseases_result_bmi <- diseases_result_bmi[diseases_result_bmi %in% colnames(disease)[-1]]
diseases_result_bmi <- data.frame(PHENO = rep(diseases_result_bmi, each = 4),
                              MODEL = rep(c("DUP", "DEL", "M", "U"), length(diseases_result_bmi)),
                              MEDIATOR = "BMI")

# Loop over diseases
for (p in unique(diseases_result_bmi$PHENO)) {
  
  print(paste0("Analyzing: ", p))
  
  # Loop over models
  for (m in c("DUP", "DEL", "M", "U")) {
    
    print(m)
    
    # Subset data (phenotype & model of interest)
    df <- left_join(na.omit(disease[, c("IID", p)]), cnv_geno[, c("IID", m)], by = "IID")
    colnames(df) <- c("IID", "PHENO", "CNV") 
    
    # Add covariates & BMI
    df <- left_join(df, bmi, by = "IID")
    df <- left_join(df, cov, by = "IID")
    
    # Fit model (Firth regression; do not include sex as a covariate for sex-specific diseases)
    if (p %in% c("PC","BC","OC","endometriosis","menstruation")) {
      fit <- logistf(PHENO ~ ., data = df[-c(1,6)], plconf = 2, maxit = 100, maxstep = 10)
    } else {
      fit <- logistf(PHENO ~ ., data = df[-c(1)], plconf = 2, maxit = 100, maxstep = 10)
    }
    
    # Fill in summary table
    diseases_result_bmi[which(diseases_result_bmi$PHENO == p & diseases_result_bmi$MODEL == m), "BETA"] <- fit$coefficients[2]
    diseases_result_bmi[which(diseases_result_bmi$PHENO == p & diseases_result_bmi$MODEL == m), "SE"] <- sqrt(diag(vcov(fit)))[2]
    diseases_result_bmi[which(diseases_result_bmi$PHENO == p & diseases_result_bmi$MODEL == m), "P"] <- fit$prob[2]
    diseases_result_bmi[which(diseases_result_bmi$PHENO == p & diseases_result_bmi$MODEL == m), "L95"] <- fit$ci.lower[2]
    diseases_result_bmi[which(diseases_result_bmi$PHENO == p & diseases_result_bmi$MODEL == m), "U95"] <- fit$ci.upper[2]
    
  }
  
}
rm(p, m, df, fit)

# Save data
fwrite(diseases_result_bmi, "/data/correction_BMI/16p11.2_BP4-5_binaryPheWAS_firth_BMIcorrected.txt", col.name = T, row.name = F, sep = "\t", quote = F)



### EA ################################################

# Create a dataframe to store results
diseases_result_ea <- df_cor[which(df_cor$MEDIATOR == "EA"), "PHENO"]
diseases_result_ea <- diseases_result_ea[diseases_result_ea %in% colnames(disease)[-1]]
diseases_result_ea <- data.frame(PHENO = rep(diseases_result_ea, each = 4),
                                  MODEL = rep(c("DUP", "DEL", "M", "U"), length(diseases_result_ea)),
                                  MEDIATOR = "EA")

# Loop over diseases
for (p in unique(diseases_result_ea$PHENO)) {
  
  print(paste0("Analyzing: ", p))
  
  # Loop over models
  for (m in c("DUP", "DEL", "M", "U")) {
    
    print(m)
    
    # Subset data (phenotype & model of interest)
    df <- left_join(na.omit(disease[, c("IID", p)]), cnv_geno[, c("IID", m)], by = "IID")
    colnames(df) <- c("IID", "PHENO", "CNV") 
    
    # Add covariates & EA
    df <- left_join(df, ea, by = "IID")
    df <- left_join(df, cov, by = "IID")
    
    # Fit model (Firth regression; do not include sex as a covariate for sex-specific diseases)
    if (p %in% c("PC","BC","OC","endometriosis","menstruation")) {
      fit <- logistf(PHENO ~ ., data = df[-c(1,6)], plconf = 2, maxit = 100, maxstep = 10)
    } else {
      fit <- logistf(PHENO ~ ., data = df[-c(1)], plconf = 2, maxit = 100, maxstep = 10)
    }
    
    # Fill in summary table
    diseases_result_ea[which(diseases_result_ea$PHENO == p & diseases_result_ea$MODEL == m), "BETA"] <- fit$coefficients[2]
    diseases_result_ea[which(diseases_result_ea$PHENO == p & diseases_result_ea$MODEL == m), "SE"] <- sqrt(diag(vcov(fit)))[2]
    diseases_result_ea[which(diseases_result_ea$PHENO == p & diseases_result_ea$MODEL == m), "P"] <- fit$prob[2]
    diseases_result_ea[which(diseases_result_ea$PHENO == p & diseases_result_ea$MODEL == m), "L95"] <- fit$ci.lower[2]
    diseases_result_ea[which(diseases_result_ea$PHENO == p & diseases_result_ea$MODEL == m), "U95"] <- fit$ci.upper[2]
    
  }
  
}
rm(p, m, df, fit)

# Save data
fwrite(diseases_result_ea, "/data/correction_EA/16p11.2_BP4-5_binaryPheWAS_firth_EAcorrected.txt", col.name = T, row.name = F, sep = "\t", quote = F)



### HEIGHT ################################################

# Create a dataframe to store results
diseases_result_height <- df_cor[which(df_cor$MEDIATOR == "height"), "PHENO"]
diseases_result_height <- diseases_result_height[diseases_result_height %in% colnames(disease)[-1]]
diseases_result_height <- data.frame(PHENO = rep(diseases_result_height, each = 4),
                                  MODEL = rep(c("DUP", "DEL", "M", "U"), length(diseases_result_height)),
                                  MEDIATOR = "height")

# Loop over diseases
for (p in unique(diseases_result_height$PHENO)) {
  
  print(paste0("Analyzing: ", p))
  
  # Loop over models
  for (m in c("DUP", "DEL", "M", "U")) {
    
    print(m)
    
    # Subset data (phenotype & model of interest)
    df <- left_join(na.omit(disease[, c("IID", p)]), cnv_geno[, c("IID", m)], by = "IID")
    colnames(df) <- c("IID", "PHENO", "CNV") 
    
    # Add covariates & HEIGHT
    df <- left_join(df, height, by = "IID")
    df <- left_join(df, cov, by = "IID")
    
    # Fit model (Firth regression; do not include sex as a covariate for sex-specific diseases)
    if (p %in% c("PC","BC","OC","endometriosis","menstruation")) {
      fit <- logistf(PHENO ~ ., data = df[-c(1,6)], plconf = 2, maxit = 100, maxstep = 10)
    } else {
      fit <- logistf(PHENO ~ ., data = df[-c(1)], plconf = 2, maxit = 100, maxstep = 10)
    }
    
    # Fill in summary table
    diseases_result_height[which(diseases_result_height$PHENO == p & diseases_result_height$MODEL == m), "BETA"] <- fit$coefficients[2]
    diseases_result_height[which(diseases_result_height$PHENO == p & diseases_result_height$MODEL == m), "SE"] <- sqrt(diag(vcov(fit)))[2]
    diseases_result_height[which(diseases_result_height$PHENO == p & diseases_result_height$MODEL == m), "P"] <- fit$prob[2]
    diseases_result_height[which(diseases_result_height$PHENO == p & diseases_result_height$MODEL == m), "L95"] <- fit$ci.lower[2]
    diseases_result_height[which(diseases_result_height$PHENO == p & diseases_result_height$MODEL == m), "U95"] <- fit$ci.upper[2]
    
  }
  
}
rm(p, m, df, fit)

# Save data
fwrite(diseases_result_height, "/data/correction_height/16p11.2_BP4-5_binaryPheWAS_firth_HEIGHTcorrected.txt", col.name = T, row.name = F, sep = "\t", quote = F)
