==> 02_covariate_analysis/03_summary_covariate_correction.R <==
########################################################
# Libraries
########################################################
library(dplyr)
library(data.table)
library(ggplot2)
library(MetBrewer)
library(ggrepel)
library(showtext)
font_add("Arial", "/System/Library/Fonts/Supplemental/Arial.ttf", bold = "/System/Library/Fonts/Supplemental/Arial Bold.ttf", italic = "/System/Library/Fonts/Supplemental/Arial Italic.ttf", bolditalic = "/System/Library/Fonts/Supplemental/Arial Bold Italic.ttf")



########################################################
# STEP 1: Load PheWAS data
########################################################

### No correction ######################################

# Load continuous trait and disease associations (generated by script/01_PheWAS/02_PheWAS.R)
pheno_noC <- as.data.frame(fread("/data/16p11.2_BP4-5_continuousPheWAS.txt", col.names = c("PHENO", "MODEL", "BETA", "SE", "P", "L95", "U95")))
disease_noC <- as.data.frame(fread("/data/16p11.2_BP4-5_binaryPheWAS_firth.txt", select = c(1:7), col.names = c("PHENO", "MODEL", "BETA", "SE", "P", "L95", "U95")))

# Merge
df_noC <- rbind(pheno_noC, disease_noC)
rm(pheno_noC, disease_noC)


### BMI correction #####################################

# Load continuous trait and disease associations (generated by script/02_covariate_analysis/02_PheWAS_covariate_corrected.R)
pheno_BMI <- as.data.frame(fread("/data/correction_BMI/16p11.2_BP4-5_continuousPheWAS_BMIcorrected.txt", col.names = c("PHENO", "MODEL", "MEDIATOR", "BETA_C", "SE_C", "P_C", "L95_C", "U95_C")))
disease_BMI <- as.data.frame(fread("/data/correction_BMI/16p11.2_BP4-5_binaryPheWAS_firth_BMIcorrected.txt", col.names = c("PHENO", "MODEL", "MEDIATOR", "BETA_C", "SE_C", "P_C", "L95_C", "U95_C")))

# Merge
df_BMI <- rbind(pheno_BMI, disease_BMI)
rm(pheno_BMI, disease_BMI)
df_BMI <- left_join(df_noC, df_BMI, by = c("PHENO", "MODEL"))


### EA correction ######################################

# Load continuous trait and disease associations (generated by script/02_covariate_analysis/02_PheWAS_covariate_corrected.R)
pheno_EA <- as.data.frame(fread("/data/correction_EA/16p11.2_BP4-5_continuousPheWAS_EAcorrected.txt", col.names = c("PHENO", "MODEL", "MEDIATOR", "BETA_C", "SE_C", "P_C", "L95_C", "U95_C")))
disease_EA <- as.data.frame(fread("/data/correction_EA/16p11.2_BP4-5_binaryPheWAS_firth_EAcorrected.txt", col.names = c("PHENO", "MODEL", "MEDIATOR", "BETA_C", "SE_C", "P_C", "L95_C", "U95_C")))

# Merge
df_EA <- rbind(pheno_EA, disease_EA)
rm(pheno_EA, disease_EA)
df_EA <- left_join(df_noC, df_EA, by = c("PHENO", "MODEL"))


### Height correction ###################################

# Load continuous trait and disease associations (generated by script/02_covariate_analysis/02_PheWAS_covariate_corrected.R)
pheno_height <- as.data.frame(fread("/data/correction_height/16p11.2_BP4-5_continuousPheWAS_Heightcorrected.txt", col.names = c("PHENO", "MODEL", "MEDIATOR", "BETA_C", "SE_C", "P_C", "L95_C", "U95_C")))
disease_height <- as.data.frame(fread("/data/correction_height/16p11.2_BP4-5_binaryPheWAS_firth_Heightcorrected.txt", col.names = c("PHENO", "MODEL", "MEDIATOR", "BETA_C", "SE_C", "P_C", "L95_C", "U95_C")))

# Merge
df_height <- rbind(pheno_height, disease_height)
rm(pheno_height, disease_height)
df_height <- left_join(df_noC, df_height, by = c("PHENO", "MODEL"))



########################################################
# STEP 2: Merge & select data
########################################################

# Merge
df <- rbind(df_BMI, df_height, df_EA)
rm(df_BMI, df_noC, df_height, df_EA)

# Retain the most significant model (no correction)
df <- df %>% 
      group_by(PHENO, MEDIATOR) %>% 
      slice(which.min(P))
df <- na.omit(df)

# Save
fwrite(df, "/data/summary_correction.txt", col.name = T, row.name = F, sep = "\t", quote = F)

# Exclude associations that are never significant
df <- df[which(df$P <= 0.05/117 | (df$P > 0.05/117 & df$P_C <= 0.05/117)), ]



########################################################
# STEP 3: Remodel & Factorize data
########################################################

# Factorize models (shape & columns)
df$MODEL <- factor(df$MODEL, levels = rev(c("DEL", "DUP", "M", "U")), labels = rev(c("deletion-only", "duplication-only", "mirror", "U-shape")))

# Factorize mediator corrections (rows)
df$MEDIATOR <- factor(df$MEDIATOR, levels = c("BMI", "EA", "height"), labels = c("BMI", "EA", "height"))

# Factorize physiological system category (color)
df[which(df$PHENO %in% c("IDA", "B12A", "AA", "anaemia", "sarcoidosis", "WBC_count", "reticulocyte_count", "RBC_count", "platelet_count", "neutrophil_count", "monocyte_count", "MCH", "lymphocyte_count", "eosinophil_count")), "CAT"] <- "Hematologic"
df[which(df$PHENO %in% c("systolic_BP", "heart_rate", "diastolic_BP", "IHD", "aneurysm", "PE", "HTN_essential", "valves", "cardiomyopathies", "conduction", "stroke_ischemic", "stroke_hemorrhagic")), "CAT"] <- "Cardiovascular"
df[which(df$PHENO %in% c("FVC", "emphysema", "COPD", "asthma", "pneumonia")), "CAT"] <- "Pulmonary"
df[which(df$PHENO %in% c("GGT", "bilirubin", "AST", "ALT", "albumin", "total_protein", "ALP", "hepatic_fibrosis", "cholelithiasis", "bilirubin_disorder", "IBD", "celiac", "CRC")), "CAT"] <- "Hepatic & Gastrointestinal"
df[which(df$PHENO %in% c("urea", "urate", "cystatinC", "Cr", "AKI", "CKD", "KS", "PKD", "KC")), "CAT"] <- "Renal"
df[which(df$PHENO %in% c("height", "GS", "vitamin_D", "phosphate", "Ca", "BMD", "IGF1", "osteoporosis", "mineral_disorder", "OA", "RA", "gout", "SLE", "hernia")), "CAT"] <- "Musculoskeletal & Connective"
df[which(df$PHENO %in% c("cornea", "cataract", "glaucoma")), "CAT"] <- "Ophthalmic"
df[which(df$PHENO %in% c("psoriasis")), "CAT"] <- "Dermatologic"
df[which(df$PHENO %in% c("BC", "OC", "PC", "endometriosis", "menstruation", "testosterone", "SHBG", "menopause", "menarche", "facial_hair", "birth_weight_first_child", "balding")), "CAT"] <- "Sex-specific"
df[which(df$PHENO %in% c("SCZ", "bipolar", "depression", "neuroticism", "PD", "AD", "MS", "epilepsy", "headaches", "apnea", "sleep", "fluid_intelligence")), "CAT"] <- "Neuropsychiatric"
df[which(df$PHENO %in% c("CRP", "hypothyroidism", "hyperthyroidism", "pituitary", "WHRadjBMI", "WHR", "weight", "body_fat_mass", "BMR", "BMI", "birthweight", "TG", "LPA", "LDL", "HDL", "cholesterol", "ApoA", "ApoB", "lipid", "glucose", "HbA1c", "T1D")), "CAT"] <- "Metabolic & Endocrine"
df$CAT <- factor(df$CAT, levels = c("Dermatologic", "Sex-specific", "Metabolic & Endocrine", "Hematologic", "Hepatic & Gastrointestinal", "Cardiovascular", "Pulmonary", "Neuropsychiatric", "Musculoskeletal & Connective", "Ophthalmic", "Renal"))

# Set size based on significance upon correction
df$SIZE <- "large"
df[which(df$P_C > 0.05/117), "SIZE"] <- "small"
df$SIZE <- factor(df$SIZE, levels = c("small", "large"))

# Set label: lose significance after covariate adjustment (main plot)
df$LABEL_main <- df$PHENO
df[which(df$P_C <= 0.05/117), "LABEL_main"] <- NA
df[which(df$MODEL == "deletion-only" & df$MEDIATOR == "BMI"), "LABEL_main"] <- NA

# Set label: lose significance after covariate adjustment (zoom plot)
df$LABEL_zoom <- df$PHENO
df[which(df$P_C <= 0.05/117), "LABEL_zoom"] <- NA

# Set label: gain significance after covariate adjustment (supplemental plot)
df$LABEL_P <- ""
df[which(df$P > 0.05/117 & df$P_C <= 0.05/117), "LABEL_P"] <- df[which(df$P > 0.05/117 & df$P_C <= 0.05/117), "PHENO"]



########################################################
# STEP 4: Plot (only best model - no correction)
########################################################

# Effect size, beta (main plot)
p_beta <- ggplot(df, aes(x = BETA, y = BETA_C, color = CAT, shape = MODEL, size = SIZE)) +
              # Identity line
              geom_abline(intercept = 0, slope = 1, lwd = 0.5, color = "azure3", lty = "dashed") +
              # Points
              geom_point(aes(fill = CAT), alpha = 0.5) +
              scale_color_manual("Trait category", values = c(met.brewer("Signac"))[c(1:5,7,10:14)], drop = F) +
              scale_fill_manual("Trait category", values = c(met.brewer("Signac"))[c(1:5,7,10:14)], drop = F) +
              scale_shape_manual("Best model", values = rev(c(25, 24, 21, 23))) +
              scale_size_manual("", values = c(1, 2.5), guide = "none") +
              # Facet
              facet_grid(MEDIATOR ~ MODEL) +
              # Label
              geom_label_repel(aes(label = LABEL_main), size = 2.75, show.legend = F,
                              box.padding = 0.15, label.padding = 0.15, max.overlaps = Inf) +
              # Layout
              xlab(expression("effect size ("*beta*")")) + ylab(expression("effect size ("*beta["adjusted"]*")")) + 
              coord_cartesian(ylim = c(-2.4, 4), xlim = c(-2.4, 4)) +
              theme_bw() +
              theme(text = element_text(family = "Arial"),
                    axis.text.x = element_text(size = 12), 
                    axis.text.y = element_text(size = 11), 
                    axis.title.x = element_text(size = 13, margin = margin(t = 7, r = 0, b = 0, l = 0)),
                    axis.title.y = element_text(size = 13, margin = margin(t = 0, r = 7, b = 0, l = 0)),
                    legend.title = element_text(size = 11, face = "bold"), legend.text = element_text(size = 10),
                    legend.key.size = unit(1, 'lines'),
                    strip.text = element_text(size = 10, color = "#2B4050", face = "bold"),
                    strip.background = element_rect(fill = "#F4F4F4", color = "#2B4050")) +
              guides(colour = guide_legend(override.aes = list(size = 3.5)))
  
# Save plot
showtext_auto()   
p_beta
ggsave("/plot/16p11.2_BP4-5_confounding_BMI_EA_height_by_model_MAIN.pdf", width = 8.4, height = 4.4)
showtext_auto(FALSE)


# Effect size, beta (main zoom plot)
p_beta_zoom <- ggplot(df[which(df$MODEL == "deletion-only" & df$MEDIATOR == "BMI"), ], aes(x = BETA, y = BETA_C, color = CAT, shape = MODEL, size = SIZE)) +
                # Identity line
                geom_abline(intercept = 0, slope = 1, lwd = 0.5, color = "azure3", lty = "dashed") +
                # Points
                geom_point(aes(fill = CAT), alpha = 0.5, show.legend = F) +
                scale_color_manual("Trait category", values = c(met.brewer("Signac"))[c(1:5,7,10:14)], drop = F) +
                scale_fill_manual("Trait category", values = c(met.brewer("Signac"))[c(1:5,7,10:14)], drop = F) +
                scale_shape_manual("Best model", values = rev(c(25, 24, 21, 23)), drop = F) +
                scale_size_manual("", values = c(2, 4), guide = "none") +
                # Facet
                facet_grid(MEDIATOR ~ MODEL) +
                # Label
                geom_label_repel(aes(label = LABEL_zoom), size = 3, show.legend = F,
                                 box.padding = 0.15, label.padding = 0.15, 
                                 max.overlaps = Inf) +
                # Layout
                xlab(expression("effect size ("*beta*")")) + ylab(expression("effect size ("*beta["adjusted"]*")")) + 
                coord_cartesian(ylim = c(-1.5, 3), xlim = c(-1.5, 3)) +
                theme_bw() +
                theme(text = element_text(family = "Arial"),
                      axis.text.x = element_text(size = 12), 
                      axis.text.y = element_text(size = 11), 
                      axis.title.x = element_text(size = 13, margin = margin(t = 7, r = 0, b = 0, l = 0)),
                      axis.title.y = element_text(size = 13, margin = margin(t = 0, r = 7, b = 0, l = 0)),
                      legend.title = element_text(size = 11, face = "bold"), legend.text = element_text(size = 10),
                      legend.key.size = unit(1, 'lines'),
                      strip.text = element_text(size = 10, color = "#2B4050", face = "bold"),
                      strip.background = element_rect(fill = "#F4F4F4", color = "#2B4050"))

# Save plot
showtext_auto()   
p_beta_zoom
ggsave("/plot/16p11.2_BP4-5_confounding_BMI_del_ZOOM.pdf", width = 4, height = 3.5)
showtext_auto(FALSE)


# P-values (supplemental plot)
p_P <- ggplot(df, aes(x = -log10(P), y = -log10(P_C), color = CAT, shape = MODEL, size = SIZE)) +
          # Lines
          geom_abline(intercept = 0, slope = 1, lwd = 0.5, color = "azure3") +
          geom_vline(xintercept = c(-log10(0.05/117)), color = "gray45", lwd = 0.5, lty = "dashed") +
          geom_hline(yintercept = c(-log10(0.05/117)), color = "gray45", lwd = 0.5, lty = "dashed") +
          # Points
          geom_point(aes(fill = CAT), alpha = 0.5) +
          scale_color_manual("Trait category", values = c(met.brewer("Signac"))[c(1:5,7,10:14)], drop = F) +
          scale_fill_manual("Trait category", values = c(met.brewer("Signac"))[c(1:5,7,10:14)], drop = F) +
          scale_shape_manual("Best model", values = rev(c(25, 24, 21, 23)), drop = F) +
          scale_size_manual("", values = c(1, 2.5), guide = "none") +
          # Facet
          facet_grid(MEDIATOR ~ MODEL) +
          # Label
          geom_label_repel(aes(label = LABEL_P), size = 2.5, show.legend = F,
                           box.padding = 0.15, label.padding = 0.15, 
                           max.overlaps = Inf) +
          # Layout
          xlab(expression(-log["10"]*"(P)")) + ylab(expression(-log["10"]*"(P"["adjusted"]*")")) + 
          theme_bw() +
          theme(text = element_text(family = "Arial"),
                axis.text.x = element_text(size = 12), 
                axis.text.y = element_text(size = 11), 
                axis.title.x = element_text(size = 13, margin = margin(t = 7, r = 0, b = 0, l = 0)),
                axis.title.y = element_text(size = 13, margin = margin(t = 0, r = 7, b = 0, l = 0)),
                legend.title = element_text(size = 11, face = "bold"), legend.text = element_text(size = 10),
                legend.key.size = unit(1, 'lines'),
                strip.text = element_text(size = 10, color = "#2B4050", face = "bold"),
                strip.background = element_rect(fill = "#F4F4F4", color = "#2B4050")) +
          guides(colour = guide_legend(override.aes = list(size = 3.5)))
        

# Save plots
showtext_auto()   
p_P
ggsave("/plot/16p11.2_BP4-5_confounding_BMI_TDI_EA_by_model_P_VALUES.pdf", width = 8.4, height = 4.4) 
showtext_auto(FALSE)
